Refactor : 좋아요 생성, 삭제 관련 로직 추가
